Antes de começar a implementação do uso de cores através de índices

+++ src/commonMain/kotlin/vitorscoelho/gyncanvas/core/primitives/Color.kt
@@ -1,9 +1,9 @@
 package vitorscoelho.gyncanvas.core.primitives
 
 interface Color {
-    val red: Float
-    val green: Float
-    val blue: Float
+    val red: Short
+    val green: Short
+    val blue: Short
     val alpha: Float
         get() = 1f
 }
